Move v1 tvl in a separate folder
